backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-genai==1.46.0
googleapis-common-protos==1.72.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.3.0
pypdfium2==5.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
